Add jar to gradle
