--- src/main/java/bobAndSteve/exception/BobAndSteveException.java
@@ -1,7 +1,0 @@
-package bobAndSteve.exception;
-
-public class BobAndSteveException extends Exception {
-    public BobAndSteveException(String message) {
-        super(message);
-    }
-}
