--- src/main/java/bobAndSteve/exception/ListIndexOutOfBoundException.java
@@ -1,7 +1,0 @@
-package bobAndSteve.exception;
-
-public class ListIndexOutOfBoundException extends BobAndSteveException {
-    public ListIndexOutOfBoundException(String message) {
-        super(message);
-    }
-}
