+++ src/main/java/bobandsteve/command/CommandEnum.java
@@ -1,0 +1,13 @@
+package bobandsteve.command;
+
+public enum CommandEnum {
+    HELP,
+    MARK,
+    UNMARK,
+    TODO,
+    DEADLINE,
+    EVENT,
+    LIST,
+    DELETE,
+    BYE;
+}
