+++ src/main/java/bobandsteve/exception/BobAndSteveException.java
@@ -1,0 +1,7 @@
+package bobandsteve.exception;
+
+public class BobAndSteveException extends Exception {
+    public BobAndSteveException(String message) {
+        super(message);
+    }
+}
