+++ src/main/java/bobandsteve/exception/InvalidCommandException.java
@@ -1,0 +1,7 @@
+package bobandsteve.exception;
+
+public class InvalidCommandException extends BobAndSteveException {
+    public InvalidCommandException(String message) {
+        super(message);
+    }
+}
