+++ src/main/java/bobandsteve/exception/InvalidCommandFormatException.java
@@ -1,0 +1,7 @@
+package bobandsteve.exception;
+
+public class InvalidCommandFormatException extends BobAndSteveException {
+    public InvalidCommandFormatException(String message) {
+        super(message);
+    }
+}
