+++ src/main/java/bobandsteve/exception/ListIndexOutOfBoundException.java
@@ -1,0 +1,7 @@
+package bobandsteve.exception;
+
+public class ListIndexOutOfBoundException extends BobAndSteveException {
+    public ListIndexOutOfBoundException(String message) {
+        super(message);
+    }
+}
